moved checking to service - didn't make frames come faster
